+++ CHANGES.md
@@ -22,6 +22,7 @@
 * Remove `io.l5d.adminMetricsExport` as a telemeter plugin, it is now always
   loaded
 * Namerd Admin now works with Dtabs of arbitrary size
+* Fix query parameter decoding when rewriting proxied requests
 
 ## 0.8.6 2017-01-19